requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
anthropic==0.43.0
elevenlabs==1.50.0
openai-whisper==20240930
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
soundfile==0.13.1
numpy==2.2.1
torch==2.5.1
